==> parameters_processing/01_parameter_estimation.Rmd <==
---
title: "Parameter estimation pipeline from cleaned Licor data"
author: "Jessica Guo, Tonantzin Tarin"
output: github_document
urlcolor: blue
editor_options: 
  chunk_output_type: console
---

Load necessary packages and functions. 
```{r}
library(dplyr)
source("R/C4_Collatz.R")
source("R/LRC.R")
source("R/Gs2.R")# Best option determined from 'compare_Gs_methods.Rmd'
source("R/Rd.R")
# devtools::install("~/plantecophys")
```


When data from a new experiment becomes available, record the experiment metadata into a local copy of "experiments.csv" and upload to the "cleaned_data" folder. Columns include ID, Date, Genotype, Treatment, A_Ci, and A_Qin. ID and Date are in YYYYMMDD format and are synonymous unless multiple experiments were completed on the same date (e.g., outdoor experiments). ID should be read in as a character. A_Ci and A_Qin columns are binary 1/0 to indicate whether each type of curve is available. 

Read in the experiment metadata. 
```{r}
expDF <- read.csv("../../sentinel-detection/data/cleaned_data/experiments.csv",
                  colClasses = c("character", "numeric", rep("character",2), rep("numeric",2)))
```

Because new experiments are conducted periodically, load the previously estimated parameter sheet. Only new  experiments (without existing parameters data) will be run. 
```{r}
params <- read.csv("../../sentinel-detection/data/derived_data/parameters_data.csv")
```

First, run the C4_Collatz.R function on ACi data to estimate vmax, or the maximum rate of carboxylation (umol CO2 m^-2 s^-1) for each plant replicate. This function uses a Bayesian estimation routine on a C4 photosynthesis model (Collatz et al. 1992). 
```{r}
aci_ID <- expDF$ID[which(expDF$A_Ci == 1 & expDF$Date > max(params$Date))] 
for(i in 1:length(aci_ID)){
  C4_Collatz(fileID = aci_ID[i])
  print(paste0(aci_ID[i], " completed"))
}
```

Second, run the LRC.R function on AQ data to estimate a suite of 6 light-response parameters for each plant replicate.  Here, the nonrectangular hyperbola model of Marshall & Biscoe (1980) is used to estiamte Am (gross maximum assimilation, umol CO2 m^-2 s^-1), AQY (apparent quantum yield umol CO2 mmol^-1 photons), Rd (dark respiration, umol CO2 m^-2, s^-1), and theta_lrc (curvature parameter, unitless). A root function is then used to find the LCPT (light compensation point, umol photons m^-2 s^-1), or PAR when assimilation is zero, and LSP (light saturation point, umol photons m^-2 s^-1), here defined as PAR when assimilation is 75% of the maximum. 
```{r}
aq_ID <- expDF$ID[which(expDF$A_Qin == 1 & expDF$Date > max(params$Date))]
for(i in 1:length(aq_ID)){
  LRC(fileID = aq_ID[i])
  print(paste0(aq_ID[i], " completed"))
}
```

Next, run the Gs2.R function on both the ACi and AQ data to obtain the stomatal sensitivity (g1BB & g1M & g1L) and cuticular conductance (g0BB & g0M & g0L, mol H2O m^-2 s^-1) across the plant replicates. Here, the data from both types of curves used across the entire range of CO2 and PAR conditions excepting extremely low values (PAR < LCPT and CO2 < 45 ppm)
```{r}
ID <- expDF$ID[which(expDF$Date > max(params$Date))]
for(i in 1:length(ID)){
  Gs(fileID = ID[i])
  print(paste0(ID[i], " completed"))
}
```

In the case of the 2020 field measurements, only ACi curves and dark-adapted respiration were measured outdoors. Two leaves per plant replicate were covered in foil for ~20 minutes prior to measurement. Due to low sample size, population-level sample means of Rd were caluclated using Rd.R. 
```{r}
ID <- expDF$ID[which(expDF$Rd == 1 & expDF$Date > max(params$Date))]
for(i in 1:length(ID)){
  Rd(fileID = ID[i])
  print(paste0(ID[i], " completed"))
}
```

Finally, once all three or four functions above are run and output parameter files produced, they can be collated and combined with the experiment data to produce a final 'parameters_data.csv'. 
```{r}
# Declare empty dataframe
params <- data.frame()

# Create vector of filenames
fnames <- c(paste0("~/sentinel-detection/data/derived_data/ACi/", list.files("~/sentinel-detection/data/derived_data/ACi", pattern = "csv")),
            paste0("~/sentinel-detection/data/derived_data/AQ/", list.files("~/sentinel-detection/data/derived_data/AQ", pattern = "csv")), 
            paste0("~/sentinel-detection/data/derived_data/stomatal/", list.files("~/sentinel-detection/data/derived_data/stomatal", pattern = "csv")),
            paste0("~/sentinel-detection/data/derived_data/Rd/", list.files("~/sentinel-detection/data/derived_data/Rd", pattern = "csv")))


for(i in 1:length(fnames)){
  temp <- read.csv(fnames[i],
                   colClasses = c(rep("character", 3), rep("numeric",3), "character"))
  params <- rbind.data.frame(params, temp)
}

# Combine parameters with experiment metadata
params2 <- left_join(params, expDF, by = "ID")
write.csv(params2, file = "~/sentinel-detection/data/derived_data/parameters_data.csv", row.names = F)
```
