==> parameters_processing/02_parameter_upload.Rmd <==
---
title: "Prepare Parameter Data for Upload to BETYdb"
author: "Kristina Riemer, Jessica Guo"
output: github_document
urlcolor: blue
---

Necessary R libraries. 
```{r}
library(dplyr)
library(stringr)
library(lubridate)
library(tidyr)
library(readxl)
library(udunits2)
```

## Physiological parameters Vmax, Rd, AQY, stomatal slope, and cuticular conductance

Read in parameters data from [sentinel-detection repository](https://github.com/danforthcenter/sentinel-detection), which should be cloned into the same folder as model-vignettes repo. 
```{r}
all_parameters <- read.csv("~/sentinel-detection/data/derived_data/parameters_data.csv",
                           stringsAsFactors = F)
```

Read in previous set of parameters. 
```{r}
#read in previous data
pparams <- read.csv("upload/phys_params/phys_parameters.csv")
pparams$local_datetime <- as.POSIXct(pparams$local_datetime)
```
Clean up data to have the final columns: 

1. `citation_author`, `citation_year`, and `citation_title`: indicate the source of the data using author, year, and title from BETYdb
2. `species`: Setaria viridis, `cultivar`: ME-034
3. `site`: one of three sites at Donald Danforth Plant Science Center, either Growth Chamber, Greenhouse, or Outdoor, using BETYdb site names
4. `treatment`: specifies record's treatment using BETYdb treatment names
5. `local_datetime`: convert date into machine readable format
6.  `Vcmax`, `leaf_respiration_rate_m2`, `quantum_efficiency`, `theta`, `stomatal_slope.BB`,  `cuticular_cond`: estimated mean of physiological parameters from parameters_data.csv
7. `leafT`: leaf temperature, which is required by BETYdb
8. `n`: sample size, always set to 1
9. `SE`: estimated standard error of physiological parameters from parameters_data.csv
10. `notes`: records which row in the parameters_data.csv the record came from
11. `access_level`: 2, equivalent to Internal & Collaborators

First, manage the chamber experiment parameters:
```{r}
upload_parameters <- all_parameters %>% 
  filter(Genotype == "ME034", 
         trait %in% c("vmax", "Rd", "AQY", "theta_lc", "g1BB", "g0BB", "g1L", "g0L")) %>% 
  mutate(date = as.POSIXct(as.character(Date), format = "%Y%m%d"), 
         treatment = case_when(Treatment == "31_22_250" ~ "regular night temperature", 
                               Treatment == "31_31_250" ~ "high night temperature", 
                               Treatment == "31_22_450" ~ "high light",
                               Treatment == "greenhouse" ~ "greenhouse",
                               Treatment == "outdoor_5cm" ~ "outdoor 5 cm density",
                               Treatment == "outdoor_JollyG" ~ "outdoor JollyG soil"),
         site = case_when(Treatment == "31_22_250" ~ "Donald Danforth Plant Science Center Growth Chamber", 
                               Treatment == "31_31_250" ~ "Donald Danforth Plant Science Center Growth Chamber", 
                               Treatment == "31_22_450" ~ "Donald Danforth Plant Science Center Growth Chamber",
                               Treatment == "greenhouse" ~ "Donald Danforth Plant Science Center Greenhouse",
                               Treatment == "outdoor_5cm" ~ "Donald Danforth Plant Science Center Outdoor",
                               Treatment == "outdoor_JollyG" ~ "Donald Danforth Plant Science Center Outdoor"),
         variable = case_when(trait == "vmax" ~ "Vcmax", 
                           trait == "Rd" ~ "leaf_respiration_rate_m2",
                           trait == "AQY" ~ "quantum_efficiency",
                           trait == "theta_lc" ~ "theta", 
                           trait == "g1BB" ~ "stomatal_slope.BB",
                           trait == "g0BB" ~ "cuticular_cond",
                           trait == "g1L" ~ "stomatal_slope",
                           trait == "g0L" ~ "cuticular_cond"),
         mean = case_when(variable == "cuticular_cond" ~ round(ud.convert(Value, "mol/m2/s", "umol/m2/s"), 0),
                          TRUE ~ round(Value, 3)),
         SE = case_when(variable == "cuticular_cond" ~ round(ud.convert(SE, "mol/m2/s", "umol/m2/s"), 0),
                          TRUE ~ round(SE, 3))) %>% 
  select(-ID, -SD, -Date.run, -Genotype, -A_Ci, -A_Qin, -Rd, -Treatment, -Value, -trait) %>% 
  #spread(Trait, value) %>% 
  dplyr::rename(local_datetime = date) %>% 
  mutate(leafT = 25, 
         n = 1,
         species = "Setaria viridis",
         cultivar = "ME-034",
         citation_author = "Zhang, Gehan, LeBauer, Riemer, Tarin, Vargas",
         citation_year = "2018", 
         citation_title = "Unpublished DARPA experimental data",
         access_level = "2" 
         ) %>%
  select(citation_author, citation_year, citation_title, 
         species, cultivar, site, treatment, local_datetime, Date, rep,
         #Vcmax, leaf_respiration_rate_m2, quantum_efficiency, stomatal_slope.BB, cuticular_cond, stomatal_slope
         variable, mean, 
         leafT, n, SE, access_level) %>%
  replace(., is.na(.), "")
```

Add a unique entity comprising of date, treatement, and rep
```{r}
upload_parameters$entity <- paste0(upload_parameters$Date, "_", upload_parameters$treatment, "_",
                                   upload_parameters$rep)
```

Save cleaned physiological parameters as a .csv file in the `phys_params` subfolder of `upload_params_data`. 
```{r}
if(!dir.exists(paste0("upload/phys_params/"))){
  dir.create(paste0("upload/phys_params/"), recursive = T)
}
write.csv(upload_parameters, "upload/phys_params/phys_parameters.csv", row.names = FALSE)
```


Split dataframe, first by new data since last  upload, then by parameter. Uploading to BETYdb using the API requires one file per parameter. 
```{r}
#create new folder with date
dname <- gsub('-', '', grep(max(upload_parameters$local_datetime), pattern = "(\\D+)", value = T))

if(!dir.exists(paste0("upload/phys_params/", dname))){
  dir.create(paste0("upload/phys_params/", dname))
}
newdir <- paste0("upload/phys_params/", dname)

#previous maximum date
prevDate <- max(pparams$local_datetime)

#split dataframe for each parameter type and save each dataframe as a csv 
params <- c("Vcmax", "leaf_respiration_rate_m2", "quantum_efficiency", "theta", "stomatal_slope.BB", "cuticular_cond", "stomatal_slope")
for(p in params){
  trait_df <- upload_parameters %>%
    filter(variable == p, local_datetime > prevDate) %>%
    spread(variable, mean)
  
  # If cuticular_cond, round; if theta, remove leafT
  if (p == "cuticular_cond") {
      trait_df2 <- trait_df %>% mutate(cuticular_cond = round(cuticular_cond, 0))
    } else if (p == "theta") {
      trait_df2 <- trait_df %>% select(-leafT)
    } else {
      trait_df2 <- trait_df
    }
  
  write.csv(trait_df2, paste0(newdir, "/phys_parameters_", p, ".csv"), row.names = FALSE)
}
```

Upload that new parameter data file using the API.
```{bash}
#change to directory of latest date
x=$(ls -d ~/model-vignettes/parameters_processing/upload/phys_params/*/ | sort -r | head -n 1)
cd $x

#add all .csv files from this diretory
for file in *.csv; do
  curl -X POST --data-binary @$file    http://welsch.cyverse.org:8000/bety/api/v1/traits.csv?key=uaZWRQT44fQVzDbxOET03EZJGXiEX9yUDEDiDwe4 -H "Content-Type: text/csv"
done
```


## Specific Leaf Area

These SLA values are calculated from leaf area and leaf dry biomass. Below, the organized data are read in, filtered, and combined to produce specific leaf area in units of m2/kg. 

These are the final columns:


1. `citation_author`, `citation_year`, and `citation_title`: indicate the source of the data using author, year, and title from BETYdb
2. `species`: Setaria viridis, `cultivar`: ME-034
3. `site`: one of three sites at Donald Danforth Plant Science Center, either Growth Chamber, Greenhouse, or Outdoor, using BETYdb site names
4. `treatment`: specifies record's treatment using BETYdb treatment names
5. `local_datetime`: convert date into machine readable format
6.  `SLA`: Specific Leaf Area calculated from single-sided fresh leaf area divided by leaf dry mass, m2/kg
8. `n`: sample size, always set to 1
9. `access_level`: 2, equivalent to Internal & Collaborators


First, load in organized biomass data from the chamber and greenhouse/outdoor experiments, which are saved in the [sentinel-detection repository](https://github.com/danforthcenter/sentinel-detection); this repo should be cloned into the same folder as model-vignettes repo. 
```{r}
load("~/sentinel-detection/data/cleaned_data/biomass/chamber_biomass.Rdata")
load("~/sentinel-detection/data/cleaned_data/biomass/greenhouse_outdoor_biomass.Rdata")
```

Select leaf biomass and leaf area that correspond to the chamber treatments; only 2 treatments have data available, as biomass has was not originally measured for 31_22_250 'regular night temperature'. 
```{r}
# Biomass and leaf area for high light treatment, 31_31_250
sla_hightemp <- chamber_biomass %>% 
 filter(genotype == "ME034V-1",
        temp == "31/31", 
        light == 250,
        !is.na(leaf_DW_mg), 
        !is.na(leaf_area_cm2),
        treatment == "control" | is.na(treatment)) %>%
  mutate(sla_initial_units = leaf_area_cm2 / leaf_DW_mg, 
         leaf_DW_kg = ud.convert(leaf_DW_mg, "mg", "kg"), 
         leaf_area_m2 = ud.convert(leaf_area_cm2, "cm2", "m2"), 
         SLA = leaf_area_m2 / leaf_DW_kg,
         Treatment = "high night temperature",
         local_datetime = harvest_date) %>% 
  select(local_datetime, Treatment, SLA) 

# Biomass and leaf area for control treatment, 31_22_450
sla_control <- chamber_biomass %>% 
 filter(genotype == "ME034V-1",
        temp == "31/22", 
        light == 430,
        !is.na(leaf_DW_mg), 
        !is.na(leaf_area_cm2),
        treatment == "control" | is.na(treatment)) %>%
  mutate(sla_initial_units = leaf_area_cm2 / leaf_DW_mg, 
         leaf_DW_kg = ud.convert(leaf_DW_mg, "mg", "kg"), 
         leaf_area_m2 = ud.convert(leaf_area_cm2, "cm2", "m2"), 
         SLA = leaf_area_m2 / leaf_DW_kg,
         Treatment = "high light",
         local_datetime = harvest_date) %>% 
  select(local_datetime, Treatment, SLA) 
```

Next, select leaf biomass and leaf area that correspond to the greenhouse and outdoor treatments. To match the physiological parameters, we use data corresponding to the first greenhouse experiment (pot treatment) and the 2nd outdoor experiment (5 cm density and JollyG pots). 
```{r}
# Biomass and leaf area for greenhouse pots
sla_gh <- greenhouse_outdoor_biomass %>%
  filter(genotype == "ME034V-1",
        exp_site == "GH", 
        exp_number == 1,
        !is.na(leaf_DW_g), 
        !is.na(leaf_area_cm2),
        treatment == "pot") %>%
  mutate(sla_initial_units = leaf_area_cm2 / leaf_DW_g, 
         leaf_DW_kg = ud.convert(leaf_DW_g, "g", "kg"), 
         leaf_area_m2 = ud.convert(leaf_area_cm2, "cm2", "m2"), 
         SLA = leaf_area_m2 / leaf_DW_kg,
         Treatment = "greenhouse",
         local_datetime = harvest_date) %>% 
  select(local_datetime, Treatment, SLA) 

# Biomass and leaf area for 2nd field experiment, 5 cm
sla_5cm <- greenhouse_outdoor_biomass %>%
  filter(genotype == "ME034V-1",
        exp_site == "Field", 
        exp_number ==  2, 
        !is.na(leaf_DW_g), 
        !is.na(leaf_area_cm2),
        treatment == "5cm") %>%
  mutate(sla_initial_units = leaf_area_cm2 / leaf_DW_g, 
         leaf_DW_kg = ud.convert(leaf_DW_g, "g", "kg"), 
         leaf_area_m2 = ud.convert(leaf_area_cm2, "cm2", "m2"), 
         SLA = leaf_area_m2 / leaf_DW_kg,
         Treatment = "outdoor 5 cm density",
         local_datetime = harvest_date) %>% 
  select(local_datetime, Treatment, SLA) 

# Biomass and leaf area for 2nd field experiment, JollyG
sla_JollyG <- greenhouse_outdoor_biomass %>%
  filter(genotype == "ME034V-1",
        exp_site == "Field", 
        exp_number ==  2, 
        !is.na(leaf_DW_g), 
        !is.na(leaf_area_cm2),
        treatment == "jolly_pot") %>%
  mutate(sla_initial_units = leaf_area_cm2 / leaf_DW_g, 
         leaf_DW_kg = ud.convert(leaf_DW_g, "g", "kg"), 
         leaf_area_m2 = ud.convert(leaf_area_cm2, "cm2", "m2"), 
         SLA = leaf_area_m2 / leaf_DW_kg,
         Treatment = "outdoor JollyG soil",
         local_datetime = harvest_date) %>% 
  select(local_datetime, Treatment, SLA) 
```

Combine and format all SLA values for upload. 
```{r}
sla <- bind_rows(sla_hightemp, sla_control, sla_gh, sla_5cm, sla_JollyG) %>%
  mutate(site = case_when(Treatment == "high light" ~ "Donald Danforth Plant Science Center Growth Chamber", 
                               Treatment == "high night temperature" ~ "Donald Danforth Plant Science Center Growth Chamber",
                               Treatment == "greenhouse" ~ "Donald Danforth Plant Science Center Greenhouse",
                               Treatment == "outdoor 5 cm density" ~ "Donald Danforth Plant Science Center Outdoor",
                               Treatment == "outdoor JollyG soil" ~ "Donald Danforth Plant Science Center Outdoor"),
         species = "Setaria viridis",
         cultivar = "ME-034",
         citation_author = "Zhang, Gehan, LeBauer, Riemer, Tarin, Vargas",
         citation_year = "2018", 
         citation_title = "Unpublished DARPA experimental data",
         access_level = "2",
         SLA = round(SLA, 2),
         treatment = Treatment) %>%
  select(citation_author, citation_year, citation_title, 
         species, cultivar, site, treatment, local_datetime, 
         SLA, access_level)
```

Save cleaned SLA as a .csv file in the `SLA` subfolder of `upload`. 
```{r}
if(!dir.exists(paste0("upload/SLA/"))){
  dir.create(paste0("upload/SLA/"), recursive = T)
}

write.csv(sla, file = "upload/SLA/sla.csv", row.names = FALSE)
```

Split dataframe by new data since last upload. 
```{r}
psla <- read.csv("upload/SLA/sla.csv")
psla$local_datetime <- as.POSIXct(psla$local_datetime)
#create new folder with date
dname <- gsub('-', '', grep(max(sla$local_datetime), pattern = "(\\D+)", value = T))

if(!dir.exists(paste0("upload/SLA/", dname))){
  dir.create(paste0("upload/SLA/", dname))
}
newdir <- paste0("upload/SLA/", dname)

#previous maximum date
prevDate <- max(psla$local_datetime)

#subset new SLA values only, to be uploaded
new_sla <- sla %>%
    filter(local_datetime > prevDate)
write.csv(new_sla, paste0(newdir, "/sla.csv"), row.names = FALSE)
```

Upload that new parameter data file using the API. 
```{bash}
#change to directory of latest date
x=$(ls -d ~/model-vignettes/parameters_processing/upload/SLA/*/ | sort -r | head -n 1)
cd $x

#add all .csv files from this diretory
for file in *.csv; do
  curl -X POST --data-binary @$file    http://welsch.cyverse.org:8000/bety/api/v1/traits.csv?key=uaZWRQT44fQVzDbxOET03EZJGXiEX9yUDEDiDwe4 -H "Content-Type: text/csv"
done
```
