==> parameters_processing/03_parameter_comparison.Rmd <==
---
title: "Compare parameters uploaded to and downloaded from BETY"
---
  
This script compares the results of what is in BETY to what has been uploaded to it, in an effort to identify mistakes or duplicates. 

Data from BETY is selected with the following SQL query: `select * from traits where treatment_id in (9000000015, 9000000016, 9000000017, 9000000018, 9000000019, 9000000020)`

The results are saved as csv in the sentinel-detection/data/derived_data/BETY folder, with a name that matches the most recent date in the `updated_at` column. 

Load libraries and both sets of data
```{r}
options(scipen = 999)
library(dplyr)
library(tidyr)
library(arsenal)
```

Import the newest BETY file.  
```{r}
path = "../../sentinel-detection/data/derived_data/BETY"
filename <- max(as.numeric(substr(list.files(path = path,
                                             pattern = "csv"), 1,8)))
in_bety <- read.csv(paste0(path, "/", filename, ".csv"), header = T)
```

For comparison, import the full data file for phys_params and SLA, found in the 'upload' folder.
```{r}
in_phys <- read.csv("upload/phys_params/phys_parameters.csv",  header = T)
in_sla <- read.csv("upload/SLA/sla.csv",  header = T)
```

Convert the treatment and variable ID columns to match upload data format. 
```{r}
bety <- in_bety %>% 
  mutate(date = as.POSIXct(as.character(date), format = "%d/%m/%Y %H:%M:%S"),
         treatment = case_when(treatment_id == "9000000015" ~ "regular night temperature",
                               treatment_id == "9000000016" ~ "high night temperature",
                               treatment_id == "9000000017" ~ "high light",
                               treatment_id == "9000000018" ~ "greenhouse",
                               treatment_id == "9000000019" ~ "outdoor 5 cm density",
                               treatment_id == "9000000020" ~ "outdoor JollyG soil"),
         variable = case_when(variable_id == "7" ~ "leaf_respiration_rate_m2",
                              variable_id == "39" ~ "quantum_efficiency",
                              variable_id == "404" ~ "stomatal_slope.BB",
                              variable_id == "4" ~ "Vcmax",
                              variable_id == "2000000027" ~ "theta",
                              variable_id == "38" ~ "cuticular_cond",
                              variable_id == "15" ~ "SLA"),
         site = case_when(site_id == "9000000005" ~ "Donald Danforth Plant Science Center Greenhouse",
                          site_id == "9000000004" ~ "Donald Danforth Plant Science Center Growth Chamber",
                          site_id == "9000000006" ~ "Donald Danforth Plant Science Center Outdoor"),
         SE = stat,
         entity = entity_id) %>% 
  select(date, site, treatment, variable, mean, n, SE, entity) %>%
  arrange(site, treatment, variable, entity) %>%
  select(-date, -entity) 

bety <- bety %>%
  mutate_if(sapply(bety, is.character), as.factor) %>%
  mutate_if(sapply(bety, is.integer), as.numeric)

#reorder
bety <- bety[order(bety$site, bety$treatment, bety$variable, bety$mean),]

#remove row names
rownames(bety) <- c()
```

Create matching version of uploaded phys_params. 
```{r}
phys <- in_phys %>%
  select(site, treatment, variable, mean, SE, entity) %>%
  arrange(site, treatment, variable, entity) %>%
  mutate(n = case_when(!is.na(SE) ~ 1,
                       is.na(SE) ~ NA_real_)) %>%
  select(-entity) %>%
  select(site, treatment, variable, mean, n, SE)
```

Create matching version of uploaded sla 
```{r}
sla <- in_sla %>%
  tidyr::gather(SLA, key = "variable", value = "mean") %>%
  mutate(entity = NA, 
         SE = NA, 
         n = 1) %>%
  select(site, treatment, variable, mean, n, SE, entity) %>%
  arrange(site, treatment, variable, entity) %>%
  select(-entity) %>%
  select(site, treatment, variable, mean, n, SE)
```

Combine traits
```{r}
all <- rbind.data.frame(phys, sla)
all$variable <- factor(all$variable, 
                       levels = unique(all$variable)[c(1,2,3,7,4,5,6)])

all <- all %>%
  mutate_if(sapply(all, is.character), as.factor)

#reorder
all <- all[order(all$site, all$treatment, all$variable, all$mean),]

#remove row names
rownames(all) <- c()
```

Compare both data versions. 
```{r}
comparedf(bety, all)
```

If no differences, should return "Equal". 

The exception is when C4_Collatz.R model has been re-run for previously existing data. Bayesian MCMC yields slightly different means of vmax.
